[Feat] #16 - 로그아웃 (fcm 토큰 초기화)
