--- app/src/main/java/com/example/pillinTimeAndroid/data/remote/SignInService.kt
@@ -1,21 +1,0 @@
-package com.example.pillinTimeAndroid.data.remote
-
-import com.example.pillinTimeAndroid.data.remote.dto.TokenDTO
-import com.example.pillinTimeAndroid.data.remote.dto.request.SignInRequest
-import com.example.pillinTimeAndroid.data.remote.dto.request.SignInSmsAuthRequest
-import com.example.pillinTimeAndroid.data.remote.dto.response.SignInSmsAuthResponse
-import com.example.pillinTimeAndroid.data.remote.dto.response.base.BaseResponse
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface SignInService {
-    @POST("/api/auth")
-    suspend fun postSignIn(
-        @Body signInRequest: SignInRequest
-    ): BaseResponse<TokenDTO>
-
-    @POST("/api/auth/sms")
-    suspend fun postSmsAuth(
-        @Body signInSmsAuthRequest: SignInSmsAuthRequest
-    ): BaseResponse<SignInSmsAuthResponse>
-}
